Adding Best First Search  algorithm file
Best First Search algorithm in artificial intelligence is an algorithm used to find the shortest path from a given start node to a goal node in a graph. The algorithm works by expanding the graph nodes in increasing order of distance from the initial node until the objective node is reached.